+++ app/Facades/FormRenderer.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+class FormRenderer extends Facade
+{
+    /**
+     * {@inheritDoc}
+     */
+    protected static function getFacadeAccessor()
+    {
+        return \App\Forms\FormRenderer::class;
+    }
+}