+++ resources/views/form/default/form_widget.blade.php
@@ -1,0 +1,1 @@
+@include($compound ? 'form.default.form_widget_compound' : 'form.default.form_widget_simple')