+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file[^1].
 ### Fixed
 - Duplicate authors with links
 - Authorities Search endpoint of admin fixed
+- Multiple item images form in admin
 
 ### Added
 - Remove query string from localized urls
@@ -16,6 +17,7 @@ All notable changes to this project will be documented in this file[^1].
 - newsletter link into footer
 - Refactor spice harvester
 - Show harvest progress in admin
+- Symfony forms
 
 ### Changed
 - made models Item, Authority, Article and Collection translatable